Fix sign/unsign compare warning Remove unused buf2 from MQ2CharacterType::GETMEMBER Fix assignment in conditional warning

+++ MQ2Main/MQ2Data.cpp
@@ -1462,7 +1462,7 @@ TLO(dataAltAbility)
         {
             if ( PALTABILITY pAbility=pAltAdvManager->GetAltAbility(nAbility))
             {
-                if (pAbility->ID == GETNUMBER() )
+                if (pAbility->ID == (DWORD)GETNUMBER() )
                 {
                     Ret.Ptr=pAbility;
                     Ret.Type=pAltAbilityType;
